+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponseragas.py
@@ -59,9 +59,9 @@ class EvaluatorResponseRagas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-07-28T09:43:05.279Z"
+    created: Optional[str] = "2026-07-29T02:51:47.026Z"
 
-    updated: Optional[str] = "2026-07-28T09:43:05.279Z"
+    updated: Optional[str] = "2026-07-29T02:51:47.026Z"
 
     updated_by_id: OptionalNullable[str] = UNSET