+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KYM1Q85MFVYXXYEVC0803VWM"
+        "01KYNWJVNQW0GE45769BGEKV7E"
     )
     r"""The unique identifier of the data source"""